refactor: separate computeScore file
